% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{conrad2}
\alias{conrad2}
\title{The pedigree of Conrad II, Holy Roman Emperor}
\format{
A data frame with 22 lines with the following variables.
\describe{
\item{list("id")}{subject id (22 is Conrad II, 2 is Louis the Stammerer}
\item{list("father")}{id of the subject's father}
\item{list("mother")}{id of the subject's mother}
\item{list("sex")}{1=male, 2=female}
}
}
\description{
A pedigree with imbreding used in examples
}
\details{
See \code{vignette("Elston-Stewart")} for the name of some other individuals in the pedigree.
}
\examples{

data(conrad2)

}
\keyword{datasets}
