% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.r
\name{es.stopCluster}
\alias{es.stopCluster}
\title{Closing clusters}
\usage{
es.stopCluster(verbose = TRUE)
}
\arguments{
\item{verbose}{if \code{TRUE}, the function will display a short information message when closing clusters}
}
\description{
This function closes clusters opened by \code{Likelihood}
}
\seealso{
\code{\link{Likelihood}}
}
