% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{fams}
\alias{fams}
\title{Fifty families}
\format{
A data frame with 802 lines with the following variables.
\describe{
\item{list("fam")}{family id}
\item{list("id")}{subject id}
\item{list("father")}{id of the subject's father}
\item{list("mother")}{id of the subject's mother}
\item{list("sex")}{1=male, 2=female}
\item{list("genotype")}{coded additively. Many genotypes are unkwnon (\code{NA})}
}
}
\description{
A set 50 pedigrees
}
\examples{

data(fams)

}
\keyword{datasets}
