% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele.di.r
\docType{data}
\name{modele.di}
\alias{modele.di}
\title{Base modele for a di-allelic marker}
\format{
An object of class \code{list} of length 4.
}
\usage{
modele.di
}
\description{
Base modele for a di-allelic marker
}
\details{
\code{modele.di} is a list with components 'name' (value "diallelic"), 'trans' : a function
'trans(of, fa, mo)' giving transmission probabilities for a genotype coded as 0, 1, 2, and 
'p.pheno', a function 'p.pheno(x, g, theta)' which always return 1.

See vignette for examples.
}
\keyword{datasets}
