\name{conrad2}
\alias{conrad2}
\docType{data}
\title{ The pedigree of Conrad II, Holy Roman Emperor }
\description{
  A pedigree with imbreding used in examples
}
\usage{data(conrad2)}
\format{
  A data frame with 22 lines with the following variables.
  \describe{
    \item{\code{id}}{subject id (22 is Conrad II, 2 is Louis the Stammerer}
    \item{\code{father}}{id of the subject's father}
    \item{\code{mother}}{id of the subject's mother}
    \item{\code{sex}}{1=male, 2=female}
  }
}
\details{See \code{vignette("Elston-Stewart")} for the name of some other individuals in the pedigree.}
\examples{
data(conrad2)
}
\keyword{datasets}

