\name{es.stopCluster}
\alias{es.stopCluster}
\title{Closing clusters} 
\description{This function closes clusters opened by \code{Likelihood}}
\usage{es.stopCluster(verbose = TRUE)}
\arguments{
  \item{verbose}{if \code{TRUE}, the function will display a short information message when closing clusters}
}
\seealso{\code{\link{Likelihood}}}

